fix(AYC-000): fix sql query for skill removal
